refactor: Move "waitForServiceInitialized" method into Steam utils
